@@ -1,6 +1,7 @@
 module Raven
-  module ActiveSupportBreadcrumbs
-    class << self
+  module Breadcrumbs
+    module ActiveSupportLogger
+      class << self
         def add(name, started, _finished, _unique_id, data)
           Raven.breadcrumbs.record do |crumb|
             crumb.data = data
@@ -10,10 +11,15 @@ module Raven
         end
 
         def inject
-          ActiveSupport::Notifications.subscribe(/.*/) do |name, started, finished, unique_id, data|
+          @subscriber = ::ActiveSupport::Notifications.subscribe(/.*/) do |name, started, finished, unique_id, data|
             add(name, started, finished, unique_id, data)
           end
         end
+
+        def detach
+          ::ActiveSupport::Notifications.unsubscribe(@subscriber)
+        end
+      end
     end
   end
 end
