@@ -12,19 +12,18 @@ module Delayed
           begin
             # Forward the call to the next callback in the callback chain
             block.call(job, *args)
-
-          rescue Exception => exception
+          rescue Exception => e
             # Log error to Sentry
             extra = {
               :delayed_job => {
-                :id          => job.id,
-                :priority    => job.priority,
-                :attempts    => job.attempts,
-                :run_at      => job.run_at,
-                :locked_at   => job.locked_at,
-                :locked_by   => job.locked_by,
-                :queue       => job.queue,
-                :created_at  => job.created_at
+                :id => job.id.to_s,
+                :priority => job.priority,
+                :attempts => job.attempts,
+                :run_at => job.run_at,
+                :locked_at => job.locked_at,
+                :locked_by => job.locked_by,
+                :queue => job.queue,
+                :created_at => job.created_at
               }
             }
             # last_error can be nil
@@ -36,16 +35,16 @@ module Delayed
             if job.respond_to?('payload_object') && job.payload_object.respond_to?('job_data')
               extra[:active_job] = Raven::ContextFilter.filter(job.payload_object.job_data)
             end
-            ::Raven.capture_exception(exception,
-                                      :logger  => 'delayed_job',
-                                      :tags    => {
+            ::Raven.capture_exception(e,
+                                      :logger => 'delayed_job',
+                                      :tags => {
                                         :delayed_job_queue => job.queue,
-                                        :delayed_job_id => job.id
+                                        :delayed_job_id => job.id.to_s
                                       },
                                       :extra => extra)
 
             # Make sure we propagate the failure!
-            raise exception
+            raise e
           ensure
             ::Raven::Context.clear!
             ::Raven::BreadcrumbBuffer.clear!
