@@ -10,10 +10,13 @@ module RackTimeoutExtensions
     # Only rack-timeout 0.3.0+ provides the request environment, but we can't
     # gate this based on a gem version constant because rack-timeout does
     # not provide one.
-    { :fingerprint => ["{{ default }}", env["REQUEST_URI"]] } if defined?(env)
+    if defined?(env)
+      { :fingerprint => ["{{ default }}", env["REQUEST_URI"]] }
+    else
+      {}
+    end
   end
 end
 
-# Include is private in Ruby 1.9
-Rack::Timeout::Error.__send__(:include, RackTimeoutExtensions)
-Rack::Timeout::RequestTimeoutException.__send__(:include, RackTimeoutExtensions)
+Rack::Timeout::Error.include(RackTimeoutExtensions)
+Rack::Timeout::RequestTimeoutException.include(RackTimeoutExtensions)
