@@ -1,5 +1,5 @@
 # frozen_string_literal: true
+
 module Raven
-  # Freezing this constant breaks in 1.9.x
-  VERSION = "2.12.0" # rubocop:disable Style/MutableConstant
+  VERSION = "3.0.4"
 end
