@@ -26,6 +26,15 @@ RSpec.describe Raven::Configuration do
     expect(subject.server).to     eq("http://sentry.localdomain:3000/sentry")
   end
 
+  describe "#breadcrumbs_logger=" do
+    it "raises error when given an invalid option" do
+      expect { subject.breadcrumbs_logger = :foo }.to raise_error(
+        Raven::Error,
+        'Unsupported breadcrumbs logger. Supported loggers: [:sentry_logger, :active_support_logger]'
+      )
+    end
+  end
+
   it "doesnt accept invalid encodings" do
     expect { subject.encoding = "apple" }.to raise_error(Raven::Error, 'Unsupported encoding')
   end
@@ -133,9 +142,11 @@ RSpec.describe Raven::Configuration do
   end
 
   context 'being initialized without a release' do
+    let(:fake_root) { "/tmp/sentry/" }
+
     before do
-      allow(File).to receive(:directory?).with(".git").and_return(false)
-      allow(File).to receive(:directory?).with("/etc/heroku").and_return(false)
+      allow(File).to receive(:directory?).and_return(false)
+      allow_any_instance_of(described_class).to receive(:project_root).and_return(fake_root)
     end
 
     it 'defaults to nil' do
@@ -148,6 +159,119 @@ RSpec.describe Raven::Configuration do
       expect(subject.release).to eq('v1')
 
       ENV.delete('SENTRY_CURRENT_ENV')
+    end
+
+    context "when git is available" do
+      before do
+        allow(File).to receive(:directory?).and_return(false)
+        allow(File).to receive(:directory?).with(".git").and_return(true)
+      end
+      it 'gets release from git' do
+        allow(Raven).to receive(:`).with("git rev-parse --short HEAD 2>&1").and_return("COMMIT_SHA")
+
+        expect(subject.release).to eq('COMMIT_SHA')
+      end
+    end
+
+    context "when Capistrano is available" do
+      let(:revision) { "2019010101000" }
+
+      before do
+        Dir.mkdir(fake_root) unless Dir.exist?(fake_root)
+        File.write(filename, file_content)
+      end
+
+      after do
+        File.delete(filename)
+        Dir.delete(fake_root)
+      end
+
+      context "when the REVISION file is present" do
+        let(:filename) do
+          File.join(fake_root, "REVISION")
+        end
+        let(:file_content) { revision }
+
+        it "gets release from the REVISION file" do
+          expect(subject.release).to eq(revision)
+        end
+      end
+
+      context "when the revisions.log file is present" do
+        let(:filename) do
+          File.join(fake_root, "..", "revisions.log")
+        end
+        let(:file_content) do
+          "Branch master (at COMMIT_SHA) deployed as release #{revision} by alice"
+        end
+
+        it "gets release from the REVISION file" do
+          expect(subject.release).to eq(revision)
+        end
+      end
+    end
+
+    context "when running on heroku" do
+      before do
+        allow(File).to receive(:directory?).and_return(false)
+        allow(File).to receive(:directory?).with("/etc/heroku").and_return(true)
+      end
+
+      context "when it's on heroku ci" do
+        it "returns nil" do
+          begin
+            original_ci_val = ENV["CI"]
+            ENV["CI"] = "true"
+
+            expect(subject.release).to eq(nil)
+          ensure
+            ENV["CI"] = original_ci_val
+          end
+        end
+      end
+
+      context "when it's not on heroku ci" do
+        around do |example|
+          begin
+            original_ci_val = ENV["CI"]
+            ENV["CI"] = nil
+
+            example.run
+          ensure
+            ENV["CI"] = original_ci_val
+          end
+        end
+
+        it "returns nil + logs an warning if HEROKU_SLUG_COMMIT is not set" do
+          logger = double("logger")
+          expect(::Raven::Logger).to receive(:new).and_return(logger)
+          expect(logger).to receive(:warn).with(described_class::HEROKU_DYNO_METADATA_MESSAGE)
+
+          expect(described_class.new.release).to eq(nil)
+        end
+
+        it "returns HEROKU_SLUG_COMMIT" do
+          begin
+            ENV["HEROKU_SLUG_COMMIT"] = "REVISION"
+
+            expect(subject.release).to eq("REVISION")
+          ensure
+            ENV["HEROKU_SLUG_COMMIT"] = nil
+          end
+        end
+      end
+    end
+  end
+
+  describe "config: backtrace_cleanup_callback" do
+    it "defaults to nil" do
+      expect(subject.backtrace_cleanup_callback).to eq(nil)
+    end
+
+    it "takes a proc and store it" do
+      subject.backtrace_cleanup_callback = proc {}
+
+      expect(subject.backtrace_cleanup_callback).to be_a(Proc)
     end
   end
 
@@ -186,6 +310,7 @@ RSpec.describe Raven::Configuration do
     end
 
     it "sets the DSN in the way we expect" do
+      expect(subject.dsn).to eq("https://66260460f09b5940498e24bb7ce093a0@sentry.io/42")
       expect(subject.server).to eq("https://sentry.io")
       expect(subject.project_id).to eq("42")
       expect(subject.public_key).to eq("66260460f09b5940498e24bb7ce093a0")
