@@ -13,6 +13,14 @@ RSpec.describe "Integration tests" do
     end
   end
 
+  it "prints deprecation warning when requiring 'sentry-raven-without-integrations'" do
+    expect do
+      require "sentry-raven-without-integrations"
+    end.to output(
+      "[Deprecation Warning] Dasherized filename \"sentry-raven-without-integrations\" is deprecated and will be removed in 4.0; use \"sentry_raven_without_integrations\" instead\n" # rubocop:disable Style/LineLength
+    ).to_stderr
+  end
+
   it "posting an exception" do
     @stubs.post('sentry/api/42/store/') { [200, {}, 'ok'] }
 
@@ -43,17 +51,17 @@ RSpec.describe "Integration tests" do
   # end
 
   it "timed backoff should prevent sends" do
-    expect(@instance.client.transport).to receive(:send_event).exactly(1).times.and_raise(Faraday::Error::ConnectionFailed, "conn failed")
+    expect(@instance.client.transport).to receive(:send_event).exactly(1).times.and_raise(Faraday::ConnectionFailed, "conn failed")
     2.times { @instance.capture_exception(build_exception) }
     expect(@io.string).to match(/Failed to submit event: ZeroDivisionError: divided by 0$/)
   end
 
   it "transport failure should call transport_failure_callback" do
-    @instance.configuration.transport_failure_callback = proc { |_e| @io.puts "OK!" }
+    @instance.configuration.transport_failure_callback = proc { |_event, error| @io.puts "OK! - #{error.message}" }
 
-    expect(@instance.client.transport).to receive(:send_event).exactly(1).times.and_raise(Faraday::Error::ConnectionFailed, "conn failed")
+    expect(@instance.client.transport).to receive(:send_event).exactly(1).times.and_raise(Faraday::ConnectionFailed, "conn failed")
     @instance.capture_exception(build_exception)
-    expect(@io.string).to match(/OK!$/)
+    expect(@io.string).to match(/OK! - conn failed$/)
   end
 
   describe '#before_send' do
