@@ -41,39 +41,22 @@ RSpec.describe JSON do
     expect(JSON.parse("[123e090000000]")).to eq [+1.0 / 0.0]
   end
 
-  if RUBY_VERSION.to_f >= 2.0 # 1.9 just hangs on this.
-    it 'it raises the correct error on strings that look like incomplete objects' do
-      expect { JSON.parse("{") }.to raise_error(JSON::ParserError)
-      expect { JSON.parse("[") }.to raise_error(JSON::ParserError)
-    end
+  it 'it raises the correct error on strings that look like incomplete objects' do
+    expect { JSON.parse("{") }.to raise_error(JSON::ParserError)
+    expect { JSON.parse("[") }.to raise_error(JSON::ParserError)
+  end
 
-    it "accepts any encoding which is internally valid" do
-      expect do
-        JSON.parse(%({"example": "this is a utf8 or ASCII string"}))
-      end.not_to raise_error
+  it "accepts any encoding which is internally valid" do
+    expect do
+      JSON.parse(%({"example": "this is a utf8 or ASCII string"}))
+    end.not_to raise_error
 
-      expect do
-        JSON.parse(%({"example": "this is a utf8 or ASCII string"}).encode("utf-16"))
-      end.not_to raise_error
+    expect do
+      JSON.parse(%({"example": "this is a utf8 or ASCII string"}).encode("utf-16"))
+    end.not_to raise_error
 
-      expect do
-        JSON.parse(%({"example": "this is a utf8 or ASCII string"}).encode("US-ASCII"))
-      end.not_to raise_error
-    end
-
-    it "blows up on circular references" do
-      data = {}
-      data['data'] = data
-      data['ary'] = []
-      data['ary'].push('x' => data['ary'])
-      data['ary2'] = data['ary']
-      data['leave intact'] = { 'not a circular reference' => true }
-
-      if RUBY_PLATFORM == 'java'
-        expect { JSON.dump(data) }.to raise_error
-      else
-        expect { JSON.dump(data) }.to raise_error(SystemStackError)
-      end
-    end
+    expect do
+      JSON.parse(%({"example": "this is a utf8 or ASCII string"}).encode("US-ASCII"))
+    end.not_to raise_error
   end
 end
