@@ -1,5 +1,3 @@
-# Encoding: utf-8
-
 require 'spec_helper'
 
 RSpec.describe Raven::Processor::Cookies do
@@ -12,18 +10,18 @@ RSpec.describe Raven::Processor::Cookies do
     test_data = {
       :request => {
         :headers => {
-          "Cookie" => "_sentry-testapp_session=SlRKVnNha2Z",
+          "Cookie" => { "_sentry-testapp_session" => "SlRKVnNha2Z" },
           "AnotherHeader" => "still_here"
         },
-        :cookies => "_sentry-testapp_session=SlRKVnNha2Z",
+        :cookies => { "_sentry-testapp_session" => "SlRKVnNha2Z" },
         :some_other_data => "still_here"
       }
     }
 
     result = @processor.process(test_data)
 
-    expect(result[:request][:cookies]).to eq("********")
-    expect(result[:request][:headers]["Cookie"]).to eq("********")
+    expect(result[:request][:cookies]).to eq({ "_sentry-testapp_session" => "********" })
+    expect(result[:request][:headers]["Cookie"]).to eq({ "_sentry-testapp_session" => "********" })
     expect(result[:request][:some_other_data]).to eq("still_here")
     expect(result[:request][:headers]["AnotherHeader"]).to eq("still_here")
   end
@@ -32,19 +30,34 @@ RSpec.describe Raven::Processor::Cookies do
     test_data = {
       "request" => {
         "headers" => {
-          "Cookie" => "_sentry-testapp_session=SlRKVnNha2Z",
+          "Cookie" => { "_sentry-testapp_session" => "SlRKVnNha2Z" },
           "AnotherHeader" => "still_here"
         },
-        "cookies" => "_sentry-testapp_session=SlRKVnNha2Z",
+        "cookies" => { "_sentry-testapp_session" => "SlRKVnNha2Z" },
         "some_other_data" => "still_here"
       }
     }
 
     result = @processor.process(test_data)
 
-    expect(result["request"]["cookies"]).to eq("********")
-    expect(result["request"]["headers"]["Cookie"]).to eq("********")
+    expect(result["request"]["cookies"]).to eq({ "_sentry-testapp_session" => "********" })
+    expect(result["request"]["headers"]["Cookie"]).to eq({ "_sentry-testapp_session" => "********" })
     expect(result["request"]["some_other_data"]).to eq("still_here")
     expect(result["request"]["headers"]["AnotherHeader"]).to eq("still_here")
+  end
+
+  it 'does not fail if it runs after Processor::RemoveCircularReferences' do
+    test_data = {
+      :request => {
+        :headers => {
+          "Cookie" => Raven::Processor::RemoveCircularReferences::ELISION_STRING,
+          "AnotherHeader" => "still_here"
+        },
+        :cookies => Raven::Processor::RemoveCircularReferences::ELISION_STRING,
+        :some_other_data => "still_here"
+      }
+    }
+
+    @processor.process(test_data)
   end
 end
