@@ -5,7 +5,7 @@ RSpec.describe Raven::Transports::Stdout do
   let(:client) { Raven::Client.new(config) }
 
   it 'should write to stdout' do
-    event = JSON.generate(Raven::Event.from_message("this is an STDOUT transport test").to_hash)
+    event = JSON.generate(Raven.capture_message("this is an STDOUT transport test").to_hash)
     expect { client.send(:transport).send_event("stdout test", event) }.to output(/\"message\":\"this is an STDOUT transport test\"/).to_stdout
   end
 end
